display a single tile from tile map texture

+++ src/core/emu/Video.h
@@ -2,14 +2,14 @@
 
 class Video {
    public:
-    static constexpr size_t kTileWidth         = 8;    // pixels
-    static constexpr size_t kTileHeight        = 8;    // pixels
-    static constexpr size_t kTileDataSize      = 16;   // bytes
-    static constexpr size_t kTileDataTableSize = 256;  // tiles
-    static constexpr size_t kDecodedTileSize =
+    static constexpr unsigned int kTileWidth         = 8;    // pixels
+    static constexpr unsigned int kTileHeight        = 8;    // pixels
+    static constexpr unsigned int kTileDataSize      = 16;   // bytes
+    static constexpr unsigned int kTileDataTableSize = 256;  // tiles
+    static constexpr unsigned int kDecodedTileSize =
         kTileWidth * kTileHeight;  // bytes
-    static constexpr size_t kTileMapSize =
-        Video::kTileDataTableSize * Video::kTileDataSize; // bytes
+    static constexpr unsigned int kTileMapSize =
+        Video::kTileDataTableSize * Video::kTileDataSize;  // bytes
 
    public:
     unsigned char tileMap[Video::kTileDataTableSize][Video::kDecodedTileSize];
